Alpha Release Files

+++ src/ykt/BeYkeRYkt/LightSource/OtherListeners/RadiusHeadListener.java
@@ -1,4 +1,6 @@
-package ykt.BeYkeRYkt.LightSource.HeadLamp;
+package ykt.BeYkeRYkt.LightSource.OtherListeners;
+
+import java.util.HashMap;
 
 import org.bukkit.Location;
 import org.bukkit.Material;
@@ -8,6 +10,7 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.PlayerBedEnterEvent;
 import org.bukkit.event.player.PlayerChangedWorldEvent;
+import org.bukkit.event.player.PlayerDropItemEvent;
 import org.bukkit.event.player.PlayerKickEvent;
 import org.bukkit.event.player.PlayerMoveEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
@@ -16,9 +19,17 @@ import org.bukkit.inventory.ItemStack;
 
 import ykt.BeYkeRYkt.LightSource.LightAPI;
 import ykt.BeYkeRYkt.LightSource.LightSource;
+import ykt.BeYkeRYkt.LightSource.HeadLamp.HeadManager;
 
-public class HeadLampListener implements Listener {
+public class RadiusHeadListener implements Listener {
 
+	private static HashMap<String, Location> pLocations = new HashMap<String, Location>();
+	
+	
+	public static HashMap<String, Location >getLocations(){
+		return pLocations;
+	}
+	
 	@EventHandler
 	public void onPlayerMove(PlayerMoveEvent event) {
 		Player player = event.getPlayer();
@@ -28,21 +39,26 @@ public class HeadLampListener implements Listener {
 		// Checking player
 		if (LightSource.getInstance().getHeadPlayers().contains(player.getName())) {
 			if (item != null && HeadManager.isHeadLamp(item)) {
-				if (event.getFrom().getBlock().getLocation().getBlockX() != event.getTo().getBlock().getLocation().getBlockX()
-						|| event.getFrom().getBlock().getLocation().getBlockY() != event.getTo().getBlock().getLocation()
-								.getBlockY()
-						|| event.getFrom().getBlock().getLocation().getBlockZ() != event.getTo().getBlock().getLocation()
-								.getBlockZ()) {
+
+			      if (!pLocations.containsKey(player.getName())) {
+			            pLocations.put(player.getName(), event.getFrom());
+			        }
+				
+			      double d = pLocations.get(player.getName()).distance(event.getTo());
+			      if (d >= LightSource.getInstance().getConfig().getDouble("Radius-update")) {
 
 					LightAPI.deleteLightSource(event.getFrom().getBlock().getLocation());
 					LightAPI.createLightSource(event.getTo().getBlock().getLocation(),
-							HeadManager.getLightLevel(item));
+					HeadManager.getLightLevel(item));
+					pLocations.put(player.getName(), event.getTo());
 				}
 			} else if (item != null && !HeadManager.isHeadLamp(item)
 					|| item == null || item.getType() == Material.AIR) {
-				LightAPI.deleteLightSourceAndUpdate(event.getFrom().getBlock().getLocation());
+				LightAPI.deleteLightSourceAndUpdate(pLocations.get(player.getName()));
 				LightSource.getInstance().getHeadPlayers().remove(player.getName());
-
+				pLocations.remove(player.getName());
+				
+				
 				if (LightSource.getInstance().getConfig().getBoolean("Debug")) {
 					LightSource.getInstance().getLogger()
 							.info("Player removed from the group");
@@ -52,6 +68,10 @@ public class HeadLampListener implements Listener {
 
 		} else if (!LightSource.getInstance().getHeadPlayers().contains(player.getName())) {
 			if (item != null && HeadManager.isHeadLamp(item)) {
+			      if (!pLocations.containsKey(player.getName())) {
+			            pLocations.put(player.getName(), event.getFrom());
+			        }
+
 				LightAPI.createLightSource(player.getLocation().getBlock().getLocation(),
 						HeadManager.getLightLevel(item));
 				LightSource.getInstance().getHeadPlayers().add(player.getName());
@@ -60,7 +80,7 @@ public class HeadLampListener implements Listener {
 					LightSource.getInstance().getLogger()
 							.info("Player added to the group");
 				}
-
+			
 			}
 		}
 		}
@@ -68,74 +88,96 @@ public class HeadLampListener implements Listener {
 
 	@EventHandler
 	public void onPlayerChangeWorlds(PlayerChangedWorldEvent event) {
-		Location loc = event.getPlayer().getLocation().getBlock().getLocation();
+		if(pLocations.containsKey(event.getPlayer())) return;
+		Location loc = pLocations.get(event.getPlayer().getName());
 
 		if (LightSource.getInstance().getHeadPlayers()
 				.contains(event.getPlayer().getName())) {
 			LightAPI.deleteLightSourceAndUpdate(loc);
+			pLocations.remove(event.getPlayer().getName());
 		}
 
 	}
 
 	@EventHandler
 	public void onPlayerTeleport(PlayerTeleportEvent event) {
-		Location loc = event.getPlayer().getLocation().getBlock().getLocation();
-
+		if(pLocations.containsKey(event.getPlayer())) return;
+		Location loc = pLocations.get(event.getPlayer().getName());
 		if (LightSource.getInstance().getHeadPlayers()
 				.contains(event.getPlayer().getName())) {
 			LightAPI.deleteLightSourceAndUpdate(loc);
+			pLocations.remove(event.getPlayer().getName());
 		}
 
 	}
 
 	@EventHandler
 	public void onPlayerDeath(PlayerDeathEvent event) {
-		Location loc = event.getEntity().getLocation().getBlock().getLocation();
+		if(pLocations.containsKey(event.getEntity())) return;
+		Location loc = pLocations.get(event.getEntity().getName());
 
 		if (LightSource.getInstance().getHeadPlayers()
 				.contains(event.getEntity().getName())) {
 			LightAPI.deleteLightSourceAndUpdate(loc);
+			pLocations.remove(event.getEntity().getName());
 		}
 
 	}
 
 	@EventHandler
 	public void onPlayerQuit(PlayerQuitEvent event) {
-		Location loc = event.getPlayer().getLocation().getBlock().getLocation();
+		if(pLocations.containsKey(event.getPlayer())) return;
+		Location loc = pLocations.get(event.getPlayer().getName());
 
 		if (LightSource.getInstance().getHeadPlayers()
 				.contains(event.getPlayer().getName())) {
 			LightAPI.deleteLightSourceAndUpdate(loc);
 			LightSource.getInstance().getHeadPlayers()
 					.remove(event.getPlayer().getName());
+			pLocations.remove(event.getPlayer().getName());
 		}
 
 	}
 
 	@EventHandler
 	public void onPlayerKick(PlayerKickEvent event) {
-		Location loc = event.getPlayer().getLocation().getBlock().getLocation();
+		if(pLocations.containsKey(event.getPlayer())) return;
+		Location loc = pLocations.get(event.getPlayer().getName());
 
 		if (LightSource.getInstance().getHeadPlayers()
 				.contains(event.getPlayer().getName())) {
 			LightAPI.deleteLightSourceAndUpdate(loc);
 			LightSource.getInstance().getHeadPlayers()
 					.remove(event.getPlayer().getName());
+			pLocations.remove(event.getPlayer().getName());
 		}
 
 	}
 
 	@EventHandler
 	public void onPlayerBedEnter(PlayerBedEnterEvent event) {
-		Location loc = event.getPlayer().getLocation().getBlock().getLocation();
-
+		if(pLocations.containsKey(event.getPlayer())) return;
+		Location loc = pLocations.get(event.getPlayer().getName());
 		if (LightSource.getInstance().getHeadPlayers()
 				.contains(event.getPlayer().getName())) {
 			LightAPI.deleteLightSourceAndUpdate(loc);
 			LightSource.getInstance().getHeadPlayers()
 					.remove(event.getPlayer().getName());
+			pLocations.remove(event.getPlayer().getName());
 		}
 
 	}
 
+	@EventHandler
+	public void onPlayerDropLight(PlayerDropItemEvent event) {
+		if(pLocations.containsKey(event.getPlayer())) return;
+		Location loc = pLocations.get(event.getPlayer().getName());
+		if (LightSource.getInstance().getHeadPlayers()
+				.contains(event.getPlayer().getName())) {
+			LightAPI.deleteLightSourceAndUpdate(loc);
+			LightSource.getInstance().getHeadPlayers()
+					.remove(event.getPlayer().getName());
+			pLocations.remove(event.getPlayer().getName());
+		}
+	}
 }
